refactor: store와 reducer 분리
- 모듈화를 위한 store와 reducer 분리
