+++ client/src/common/dispatch/user-chatroom-dispatch.ts
@@ -1,5 +1,5 @@
 import { api } from '@utils/index';
-import { store } from '@store/index';
+import store from '@store/index';
 
 const getUserChatroom = async () => {
   const userChatroom = await api.getUserChatroom();
