+++ client/src/common/dispatch/user-dispatch.ts
@@ -1,5 +1,5 @@
 import { api } from '@utils/index';
-import { store } from '@store/index';
+import store from '@store/index';
 
 const getUserInfo = async () => {
   const userInfo = await api.getUserInfo();
