+++ client/src/common/reducer/rootReducer.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import userReducer from './user-reducer';
+import sidebarReducer from './sidebar-reducer';
+
+const rootReducer = combineReducers({
+  userReducer,
+  sidebarReducer
+});
+
+export default rootReducer;
