+++ client/src/components/organisms/Header/HeaderContainer.tsx
@@ -1,7 +1,8 @@
 import { connect } from 'react-redux';
 import { Header } from './Header';
 
-function mapReduxStateToReactProps(state: any) {
+function mapReduxStateToReactProps(reducer: any) {
+  const state = reducer.userReducer;
   return {
     profileUri: state.profileUri
   };
