+++ client/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { store } from '@store/index';
+import store from '@store/index';
 import App from './shared/App';
 
 ReactDOM.render(
